requirements.txt
snowflake-connector-python==3.7.0
snowflake-sqlalchemy==1.5.1
pandas==2.1.4
numpy==1.26.3
anthropic==0.40.0
streamlit==1.31.0
plotly==5.18.0
python-dotenv==1.0.0
